Reduce stack size by VTORE alignment
This is needed to prevent cases when the start address of the A53
application isn't 0x80 bytes aligned and the M7 application must
therefore be moved to the next 0x80 aligned address. The alignment will
generate an overlap between the M7 stack and the first bytes of the A53
application as for M7 we reserve a static region of 0x2000 regardless
of the alignment procedures.

Issue: ALB-8434
Signed-off-by: Ghennadi Procopciuc <ghennadi.procopciuc@nxp.com>